requirements.txt
Flask==2.1.1
Pillow==9.1.1
Werkzeug==2.2.2